pyproject.toml
[project]
name = "esdf-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "auto-py-to-exe>=2.48.1",
    "keyboard>=0.13.5",
    "pillow>=12.1.0",
    "psutil>=7.2.1",
    "pyinstaller>=6.17.0",
    "pystray>=0.19.5",
    "pywebview>=6.1",
    "pywin32>=311",
]
